pyproject.toml
[tool.poetry]
name = "ch_wifi_monitor"
version = "0.1.0"
description = "WiFi speed monitoring via ClickHouse"
authors = ["potykion <potykion@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.10.1"
clickhouse-sqlalchemy = "^0.0.10"
speedtest-cli = "^2.1"
python-dateutil = "^2.8"
delorean = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^4.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
